+++ R/FFdownload.R
@@ -5,10 +5,10 @@
 #' (especially when including daily files) takes quite a long time!
 #'
 #' @param output_file name of the .RData file to be saved (include path if necessary)
-#' @param tempdir specify if you want to keep downloaded files somewhere save. Seems to be necessary for
+#' @param tempd specify if you want to keep downloaded files somewhere save. Seems to be necessary for
 #' reproducible research as the files on the website do change from time to time
 #' @param exclude_daily excludes the daily datasets (are not downloaded) ==> speeds the process up considerably
-#' @param download set to TRUE if you actually want to download again. set to false and specify tempdir to keep processing the already downloaded files
+#' @param download set to TRUE if you actually want to download again. set to false and specify tempd to keep processing the already downloaded files
 #' @param download_only set to FALSE if you want to process all your downloaded files at once
 #' @param listsave if not NULL, the list of unzipped files is saved here (good for processing only a limited number of files through inputlist).
 #' Is written before inputlist is processed.
@@ -17,7 +17,7 @@
 #' @return RData file
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' tempf <- tempfile(fileext = ".RData"); outd <- paste0(tempdir(),"/",format(Sys.time(), "%F_%H-%M"))
 #' temptxt <- tempfile(fileext = ".txt")
 #' # Example 1: Use FFdownload to get a list of all monthly zip-files. Save that list as temptxt.
@@ -27,18 +27,18 @@
 #' # the Momentum Factor)
 #' inputlist <- c("Research_Data_Factors","Momentum_Factor","ST_Reversal_Factor","LT_Reversal_Factor")
 #' # Now process only these files if they can be matched (download only)
-#' FFdownload(exclude_daily=FALSE,tempdir=outd,download=TRUE,download_only=FALSE,
+#' FFdownload(exclude_daily=FALSE,tempd=outd,download=TRUE,download_only=FALSE,
 #' inputlist=inputlist,output_file = tempf)
 #' list.files(outd)
 #' # Then process all the downloaded files
-#' FFdownload(output_file = tempf, exclude_daily=TRUE,tempdir=outd,download=FALSE,
+#' FFdownload(output_file = tempf, exclude_daily=TRUE,tempd=outd,download=FALSE,
 #' download_only=FALSE,inputlist=inputlist)
 #' load(tempf); FFdownload$`x_F-F_Momentum_Factor`$monthly$Temp2[1:10]
 #' # Example 2: Download all non-daily files and process them
 #' # Commented out to not being tested
 #' # tempf2 <- tempfile(fileext = ".RData");
 #' # outd2<- paste0(tempdir(),"/",format(Sys.time(), "%F_%H-%M"))
-#' # FFdownload(output_file = tempf2,tempdir = outd2, exclude_daily = TRUE, download = TRUE,
+#' # FFdownload(output_file = tempf2,tempd = outd2, exclude_daily = TRUE, download = TRUE,
 #' # download_only=FALSE, listsave=temptxt)
 #' # load(tempf2)
 #' # FFdownload$x_25_Portfolios_5x5$monthly$average_value_weighted_returns
@@ -51,7 +51,7 @@
 #' @importFrom plyr mlply
 #'
 #' @export
-FFdownload <- function(output_file = "data.Rdata", tempdir=NULL, exclude_daily=FALSE, download=TRUE, download_only=FALSE, listsave=NULL, inputlist=NULL) {
+FFdownload <- function(output_file = "data.Rdata", tempd=NULL, exclude_daily=FALSE, download=TRUE, download_only=FALSE, listsave=NULL, inputlist=NULL) {
   message("Step 1: getting list of all the csv-zip-files!\n")
   URL <- "http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html"
   pg <- xml2::read_html(URL)
@@ -91,41 +91,45 @@ FFdownload <- function(output_file = "data.Rdata", tempdir=NULL, exclude_daily=F
         utils::download.file(paste0("http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/",Flinks_final[i]), paste0(temp_download,"/", Fdest),quite=TRUE)
       }
     }
-    # copy to final tempdir if wished
-    if (!is.null(tempdir)) {dir.create(tempdir,showWarnings = FALSE); file.copy(from = paste0(temp_download,"/",gsub(pattern = "ftp/","",Flinks_final)),
-                                                                                to = tempdir, recursive=TRUE)}
-  zip_files <- list.files(temp_download, full.names = TRUE, pattern = "\\.zip$", ignore.case = TRUE) # full path
-
-  lapply(zip_files, function (x) unzip(zipfile = x, exdir = temp_download))
-
-  csv_files <- list.files(temp_download, full.names = TRUE, pattern = "\\.csv$", ignore.case = TRUE) # full path
-  csv_files2 <- list.files(temp_download, full.names = FALSE, pattern = "\\.csv$", ignore.case = TRUE) # only filenames
-  if (length(grep("daily",csv_files2,ignore.case = TRUE))){
-    csv_files2_daily <- csv_files2[grep("daily",csv_files2,ignore.case = TRUE)]
-    csv_files2_nodaily <- csv_files2[-grep("daily",csv_files2,ignore.case = TRUE)]
-  } else {csv_files2_daily <- NULL; csv_files2_nodaily <- csv_files2}
-
-  vars_nodaily <- paste0("x_", gsub("(.*)\\..*", "\\1", csv_files2_nodaily)  )
-  vars_daily <- paste0("x_", gsub("(.*)\\..*", "\\1", csv_files2_daily)  )
-  vars <- paste0("x_", gsub("(.*)\\..*", "\\1", csv_files2)  )
-
+    # copy to final tempd if wished
+    if (!is.null(tempd)) {dir.create(tempd,showWarnings = FALSE); file.copy(from = paste0(temp_download,"/",gsub(pattern = "ftp/","",Flinks_final)),
+                                                                                to = tempd, recursive=TRUE)}
+  }
   # if download_only==TRUE exit
   if(!download_only){
-    message("Step 3: Start processing ",length(Flinks_final)," csv-files\n")
-    FFdownload <- plyr::mlply(function(y) converter(y), .data=csv_files, .progress = "text")
-    names(FFdownload) <- vars
+    if (!download&is.null(tempd)) {stop("No directory given for reading files!")
+      } else if (download&is.null(tempd))  {
+        tempd <- temp_download
+      }
+    zip_files <- list.files(tempd, full.names = TRUE, pattern = "\\.zip$", ignore.case = TRUE) # full path
+    lapply(zip_files, function (x) unzip(zipfile = x, exdir = tempd))
 
+    csv_files <- list.files(tempd, full.names = TRUE, pattern = "\\.csv$", ignore.case = TRUE) # full path
+    csv_files2 <- list.files(tempd, full.names = FALSE, pattern = "\\.csv$", ignore.case = TRUE) # only filenames
+    if (length(grep("daily",csv_files2,ignore.case = TRUE))){
+      csv_files2_daily <- csv_files2[grep("daily",csv_files2,ignore.case = TRUE)]
+      csv_files2_nodaily <- csv_files2[-grep("daily",csv_files2,ignore.case = TRUE)]
+    } else {csv_files2_daily <- NULL; csv_files2_nodaily <- csv_files2}
+
+    vars_nodaily <- paste0("x_", gsub("(.*)\\..*", "\\1", csv_files2_nodaily)  )
+    vars_daily <- paste0("x_", gsub("(.*)\\..*", "\\1", csv_files2_daily)  )
+    vars <- paste0("x_", gsub("(.*)\\..*", "\\1", csv_files2)  )
+
+    message("Step 3: Start processing ",length(Flinks_final)," csv-files\n")
+    FFdata <- plyr::mlply(function(y) converter(y), .data=csv_files, .progress = "text")
+    names(FFdata) <- vars
 
     # recombine lists
     if(!exclude_daily){
       for (i in 1:length(vars_nodaily)){
         if (any(grepl(vars_nodaily[i],vars_daily))){
-          FFdownload[[eval(vars_nodaily[i])]]$daily <- FFdownload[[eval(vars_daily[grep(vars_nodaily[i],vars_daily)])]]$daily
-          FFdownload[[eval(vars_daily[grep(vars_nodaily[i],vars_daily)])]] <- NULL
+          FFdata[[eval(vars_nodaily[i])]]$daily <- FFdata[[eval(vars_daily[grep(vars_nodaily[i],vars_daily)])]]$daily
+          FFdata[[eval(vars_daily[grep(vars_nodaily[i],vars_daily)])]] <- NULL
         }
       }
     }
-    save(FFdownload, file = output_file)
+    save(FFdata, file = output_file)
+    message("Be aware that as of version 1.0.6 the saved object is named FFdata rather than FFdownload to not be confused with the corresponding command!")
   }
-  }
+
 }